// src/api/models/PatchedPersonUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PatchedPersonUpdate = {
    readonly person_id?: number;
    social_name?: string | null;
    birth_datetime?: string | null;
    profile_picture?: string | null;
    year_of_birth?: number | null;
    gender_concept?: number | null;
    ethnicity_concept?: number | null;
    race_concept?: number | null;
    location?: number | null;
};